requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
python-dotenv
google-generativeai
Flask-WTF